Move custom archived reason/explanation fields onto Barrier model
