+++ api/barriers/models.py
@@ -26,6 +26,7 @@ from api.metadata.models import BarrierType, BarrierPriority
 from api.barriers import validators
 from api.barriers.report_stages import REPORT_CONDITIONS, report_stage_status
 from api.barriers.utils import random_barrier_reference
+from api.metadata.constants import BARRIER_ARCHIVED_REASON
 
 MAX_LENGTH = settings.CHAR_FIELD_MAX_LENGTH
 
@@ -173,6 +174,10 @@ class BarrierInstance(BaseModel, FullyArchivableModel):
         through="BarrierReportStage",
         help_text="Store reporting stages before submitting",
     )
+    archived_reason = models.CharField(
+        choices=BARRIER_ARCHIVED_REASON, max_length=25, null=True
+    )
+    archived_explanation = models.TextField(blank=True, null=True)
 
     history = HistoricalRecords()
 
@@ -227,6 +232,13 @@ class BarrierInstance(BaseModel, FullyArchivableModel):
     @property
     def modified_user(self):
         return self._cleansed_username(self.modified_by)
+
+    def archive(self, user, reason=None, explanation=None):
+        self.archived_explanation = explanation
+        self.unarchived_by = None
+        self.unarchived_on = None
+        self.unarchived_reason = ""
+        super().archive(user, reason)
 
     def save(
         self, force_insert=False, force_update=False, using=None, update_fields=None
