+++ api/core/models.py
@@ -6,7 +6,6 @@ from django.db import models
 from django.utils import timezone
 
 from api.core.utils import cleansed_username
-from api.metadata.constants import ARCHIVED_REASON
 
 
 class BaseModel(models.Model):
@@ -83,11 +82,7 @@ class ArchivableModel(models.Model):
 
 class FullyArchivableModel(ArchivableModel):
     """ Archivable model with extra fields for explanation and unarchiving."""
-
-    archived_reason = models.CharField(
-        choices=ARCHIVED_REASON, max_length=25, null=True
-    )
-    archived_explanation = models.TextField(blank=True, null=True)
+    """ Archivable mixin with extra fields for unarchiving."""
 
     unarchived_reason = models.TextField(blank=True, null=True)
     unarchived_on = models.DateTimeField(blank=True, null=True)
@@ -101,10 +96,6 @@ class FullyArchivableModel(ArchivableModel):
 
     class Meta:
         abstract = True
-
-    def archive(self, user, reason=None, explanation=None):
-        self.archived_explanation = explanation
-        super().archive(user, reason)
 
     def unarchive(self, user, reason=None):
         self.unarchived_by = user
