+++ api/metadata/constants.py
@@ -96,7 +96,7 @@ ASSESMENT_IMPACT = Choices(
     ("LOW", "Low"),
 )
 
-ARCHIVED_REASON = Choices(
+BARRIER_ARCHIVED_REASON = Choices(
     ("DUPLICATE", "Duplicate"),
     ("NOT_A_BARRIER", "Not a barrier"),
     ("OTHER", "Other"),
